requirements.txt
pytest
pytest-xdist
selenium
allure-pytest
webdriver-manager
requests
faker
Appium-Python-Client
argparse
json
pymysql
flask
sqlalchemy